endopath/endoserver/requirements.txt
fastapi
uvicorn[standard]
python-multipart
bcrypt
python-dotenv
pydantic
pydantic-settings
jinja2
weasyprint
Pillow
numpy==1.26.4
tensorflow==2.15.*
matplotlib
